remove agent.call.req field
in order to open the pearly gates of cloud events we need to rip the http
request specific stuff out of the agent's plumbing. this makes less of an
improvement than maybe I would've liked, but at least that part of it is
accomplished. agent.call.RequestBody is the 'standard' way to get a handle on
the request body now, mostly due to lb_agent. this replaces the previous logic
for caching the request body for a call to use when a placer retries with
something that doesn't require an extra goroutine, this should reduce latency
through the lb agent and is in some ways simpler (and others not).
additionally, this requires Submit and GetCall to take a context argument,
which is now plumbed too.